HW ООП наследование, static выполнена
