+++ src/OOPvvedenie/transport/Bus.java
@@ -1,0 +1,8 @@
+package OOPvvedenie.transport;
+
+public class Bus extends Transport {
+
+    public Bus(String brand, String model, int productionYear, String productionCountry, String color, int maxSpeed) {
+        super(brand, model, productionYear, productionCountry, color, maxSpeed);
+    }
+}
