+++ src/OOPvvedenie/transport/Transport.java
@@ -1,0 +1,93 @@
+package OOPvvedenie.transport;
+
+public class Transport {
+    private final String brand;
+    private final String model;
+    private final int productionYear;
+    private final String productionCountry;
+    private String color;
+    private int maxSpeed;
+
+    public Transport(String brand, String model, int productionYear, String productionCountry, String color,
+                     int maxSpeed) {
+
+        if (brand == null) {
+            this.brand = "Бренд неизвестен";
+        } else {
+            this.brand = brand;
+        }
+
+        if (model == null) {
+            this.model = "Модель неизвестна";
+        } else {
+            this.model = model;
+        }
+
+        if (productionYear <= 0) {
+            this.productionYear = 2000;
+        } else {
+            this.productionYear = productionYear;
+        }
+
+        if (productionCountry == null || productionCountry.isEmpty()) {
+            this.productionCountry = "страна производитель отсутствует";
+        } else {
+            this.productionCountry = productionCountry;
+        }
+
+        setColor(color);
+        setMaxSpeed(maxSpeed);
+    }
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    public int getProductionYear() {
+        return productionYear;
+    }
+
+    public String getProductionCountry() {
+        return productionCountry;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public void setColor(String color) {
+        if (color == null || color.isEmpty()) {
+            this.color = "white";
+        } else {
+            this.color = color;
+        }
+    }
+
+    public int getMaxSpeed() {
+        return maxSpeed;
+    }
+
+    public void setMaxSpeed(int maxSpeed) {
+        if (maxSpeed <= 0 || maxSpeed > 1000) {
+            this.maxSpeed = 250;
+        } else {
+            this.maxSpeed = maxSpeed;
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "Transport{" +
+                "brand='" + brand + '\'' +
+                ", model='" + model + '\'' +
+                ", productionYear=" + productionYear +
+                ", productionCountry='" + productionCountry + '\'' +
+                ", color='" + color + '\'' +
+                ", maxSpeed=" + maxSpeed +
+                '}';
+    }
+}
